+++ src/main/java/rentalstore/MovieType.java
@@ -3,7 +3,7 @@ package rentalstore;
 public abstract class MovieType {
     public abstract double getAmount(int dayRented);
 
-    public int updateFrequentRenterPointers(int frequentRenterPoints, int dayRented){
+    public double updateFrequentRenterPointers(double frequentRenterPoints, int dayRented){
         return ++frequentRenterPoints;
     }
 }
